WIP: Separated the code into multiple files to clean it up
