+++ hardware/kicad/EveryDayCalendarSingleMonth.kicad_prl
@@ -1,6 +1,6 @@
 {
   "board": {
-    "active_layer": 34,
+    "active_layer": 32,
     "active_layer_preset": "",
     "auto_track_width": true,
     "hidden_nets": [],
@@ -62,7 +62,7 @@
       35,
       36
     ],
-    "visible_layers": "800bfaf_7fffffff",
+    "visible_layers": "800ffeb_ffffffff",
     "zone_display_mode": 1
   },
   "meta": {
